Use 4-byte card-generated salt.
The Tesla Model 3 Key Fob and Tesla NFC Phone Key both set the first
four bytes of the challenge to a random value before encrypting. This
seems like a good idea, so we are implementing that here, too.

This may also relax some export-control requirements, since without
the salt it would be straightforward to use these credentials for
strong encryption and decryption.