Avoid wearing out flash storage (if possible)
This change is an attempt to avoid unnecessary writes to
flash memory during the normal course of using the applet—
by preferring `TYPE_AES_TRANSIENT_DESELECT` and
`TYPE_AES_TRANSIENT_RESET` over the plain `TYPE_AES`. If
neither of the two "transient" types are supported then we
will fall back to using `TYPE_AES`. We are also now only
initializing the `KeyAgreement` instance once, which
should also avoid writing to flash.

Both of these changes should improve the longevity of the
key card when run on supported hardware.